frontend/requirements.txt
streamlit
pillow
pandas
matplotlib
scikit-learn
seaborn
